+++ demo/extras/src/main/java/it/gmariotti/cardslib/demo/extras/stickylist/StickyCardListView.java
@@ -39,6 +39,7 @@ import it.gmariotti.cardslib.library.view.base.CardViewWrapper;
 import se.emilsjolander.stickylistheaders.StickyListHeadersAdapter;
 import se.emilsjolander.stickylistheaders.StickyListHeadersListView;
 
+
 /**
  * @author Gabriele Mariotti (gabri.mariotti@gmail.com)
  */